--- src/components/progress/index.js
@@ -1,3 +1,0 @@
-export {
-  EuiProgress,
-} from './progress';
