+++ src/components/progress/index.ts
@@ -1,0 +1,1 @@
+export { EuiProgress } from './progress';
